rename k8s provisioner to plain-v0
