+++ provisioner/README.md
@@ -2,5 +2,5 @@
 
 The following subdirectories contain concrete Provisioner implementations.
 
-- [k8s](k8s/README.md) - provisions "plain" k8s bundles
-- [registryv1](registryv1/README.md) - provisions "registry+v1" OLM bundles
+- [plain-v0](plain-v0/README.md) - provisions "plain+v0" k8s bundles
+- [registry-v1](registry-v1/README.md) - provisions "registry+v1" OLM bundles
