+++ provisioner/plain-v0/README.md
@@ -1,3 +1,3 @@
-# k8s
+# plain-v0 
 
 A concrete Provisioner implementation for "plain" k8s bundles.
