+++ provisioner/plain-v0/manifests/01_cluster_role.yaml
@@ -1,7 +1,7 @@
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
-  name: k8s-provisioner-admin
+  name: plain-v0-provisioner-admin
 rules:
 - apiGroups: ["core.rukpak.io"]
   resources: ["*"]
