+++ provisioner/plain-v0/manifests/02_role.yaml
@@ -1,7 +1,7 @@
 apiVersion: rbac.authorization.k8s.io/v1
 kind: Role
 metadata:
-  name: k8s-provisioner-admin
+  name: plain-v0-provisioner-admin
   namespace: rukpak-system
 rules:
 - apiGroups: [""]
