+++ provisioner/plain-v0/manifests/03_serviceaccount.yaml
@@ -1,5 +1,5 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: k8s-provisioner-admin
+  name: plain-v0-provisioner-admin
   namespace: rukpak-system
