+++ provisioner/plain-v0/manifests/04_cluster_role_binding.yaml
@@ -1,13 +1,13 @@
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
 metadata:
-  name: k8s-provisioner-admin
+  name: plain-v0-provisioner-admin
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: k8s-provisioner-admin
+  name: plain-v0-provisioner-admin
 subjects:
   - apiGroup: ""
     kind: ServiceAccount
-    name: k8s-provisioner-admin
+    name: plain-v0-provisioner-admin
     namespace: rukpak-system
