+++ provisioner/plain-v0/manifests/05_role_binding.yaml
@@ -1,14 +1,14 @@
 apiVersion: rbac.authorization.k8s.io/v1
 kind: RoleBinding
 metadata:
-  name: k8s-provisioner-admin
+  name: plain-v0-provisioner-admin
   namespace: rukpak-system
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: Role
-  name: k8s-provisioner-admin
+  name: plain-v0-provisioner-admin
 subjects:
   - apiGroup: ""
     kind: ServiceAccount
-    name: k8s-provisioner-admin
-    namespace: k8s-provisioner
+    name: plain-v0-provisioner-admin
+    namespace: plain-v0-provisioner
