+++ provisioner/registry-v1/README.md
@@ -1,3 +1,3 @@
-# registryv1
+# registry-v1
 
 This provisioner was removed in commit 76fefcfcd1408d04b4eb7eb368d3eaab49565237.
